atomicClient: Check API availability
Cockpit has the following convention: When a plugin is loaded, its
body should be initially hidden.  Once the plugin is done with
initialization, it should show the body.

Cockpit is reacting to this and shows a spinner during the plugin
initialization, for example.

Another effect of this is that the user can not interact with the page
until it is initialzed.  Here, the user can not click on the "Run"
button until the proxy is ready and fully populated.

Hitting the "Run" button before the proxy is ready would lead to
exceptions like "proxy.Version is not a function", and we avoid that
by hiding the page until the proxy is ready.

Additionally, the new code here checks the "valid" field of the proxy.
It will be false if the service/object/interface that the proxy
represents isn't actually there.  In that case, we show an error
instead of the UI.  Just as with a not-yet-ready proxy, a invalid
proxy wont have any of the expected methods like proxy.Version.