storage: Don't crash while removing devices when there is no volume group
When the storage pool is not in a volume group, we would previously
try to list the physical volumes of a volume group called "".  This
would crash.

Now, if the storage pool is not in a volume group, we treat that as a
volume group without any physical or logical volumes, which allows the
rest of the code to cope with that case cleanly.